feat: support custom rate-limiting strategies (#116)
Implements support for custom rate-limiting strategies, allowing for changing the
default library behavior of waiting until the current period has expired. This allows
for significantly more flexibility in how rate-limiting is handled.

This also adds a new implementation of this strategy that simply throws an error
when being rate-limited.

One of the primary desired use cases for this is pooling auth to automatically
rotate between users whenever one user gets rate-limited. This will require a
more significant refactor of the scraper and so is not included at this time. As a
workaround, library consumers can leverage `ErrorRateLimitStrategy` to throw
immediately when being rate-limited, and use that to pool `Scraper` instances
themselves. Given that the code has a strong assumption on only being
authenticated with one user at a time, that may well be how this library
implements pooling in the future (maybe a `PooledScraper` with the same
public interface that internally manages multiple `Scraper` instances).